Refactoring code

+++ composeApp/src/commonMain/kotlin/domain/usecase/TimeFromLastRequestUseCase.kt
@@ -1,5 +1,5 @@
 package domain.usecase
 
 interface TimeFromLastRequestUseCase {
-    suspend operator fun invoke(): Double
+    suspend operator fun invoke(): Long
 }
